Make EnclosureListView a subtype of GtkBox
In ItemView, it is not clear that EnclosureListView is not leaked.
By making EnclosureListView a widget, we no longer need
enclosure_list_view_get_widget and callbacks to self-destruct the
EnclosureListView when the widget is destroyed.

Also, as we don't plan to create subtypes of EnclosureListView, we
don't need its structure to be defined "publicly" in the header, and we
can avoid having a separate private structure, which simplifies the code.