Media Player plugins improvements
This makes LifereaMediaPlayer a subtype of GtkBox. The
LifereaMediaPlayer instance now owns the PeasExtensionSet of Media
Player extensions. The LifereaMediaPlayer is set as the "plugins-box"
property of LifereaMediaPlayerActivatable at the time of construction.
Plugins just need to add themselves to that GtkBox, they don't depend on
the implementation of EnclosureListView. The plugins still have to
manage showing, hiding and destroying their widget.

The mime type check is moved to the plugins : this lets plugins handle
other MIME types than audio and video. Plugins don't even have to
provide a widget, so the interface can be used for more generic
enclosure plugins, like a custom enclosure downloader.

Enabling a plugin will now immediately call its load function.

The media player plugin now removes its widget when deactivated.