Moves media player widget out of EnclosureView
The widget almost stays in the same position. If I try to attach it
above of the enclosure view widget, it is not visible, for some reason.
The LifereaMediaPlayerActivatable plugin interface is modified to let
the plugins know when the user moves away from the item, so that the
plugin can hide its widget if appropriate.

The media player plugin is modified to keep playing when navigating
away. If a visited item contains audio while the player is already
playing, the user should skip forward to play it, as the played item is
kept at the start of the player's list.  To stop playing a track ...
well there is no stop button, [they are useless] [1], so you HAVE to
listen to the end :p

[1]: https://github.com/lwindolf/liferea/pull/191